+++ R/utils.R
@@ -1,5 +1,5 @@
 # @staticimports pkg:isstatic
-# is_all_null
+# is_all_null is_dir has_fileext
 
 # @staticimports pkg:staticimports
 #  map2