+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Update `cli_quiet()` to work as a helper function within other functions to set the cli.default_handler option to `suppressMessages()`.
 * Refactor `cli_abort_ifnot()` and similar functions to use the new `cli_if()` function.
+* Refactor `cli_list_files()` including replacing message parameter with text and allowing option to pass character vector with files directly.
 * Add `standalone-cliExtras.R` script that can be imported with `usethis::use_standalone()` (available in the development version of {usethis}).
 * Adopt lifecycle to flag `cli_paths()` as a superseded function.